refactor: item convert logic을 ItemConvertResolver 객체로 분리
